+++ Cargo.toml
@@ -85,7 +85,7 @@ smol = "1.3.0"
 smolscale = "0.3.52"
 smol-timeout = "0.6.0"
 stdcode = "0.1.14"
-structopt = "0.3.26"
+
 tap = "1.0.1"
 # tracing = "0.1.34"
 #themelio-stf = { path = "../themelio-stf" }
@@ -121,6 +121,7 @@ rusqlite = {version="0.28.0", features=["bundled"]}
 scopeguard = "1.1.0"
 base64 = "0.21.0"
 melblkidx = { git = "https://github.com/themeliolabs/melblkidx", branch = "master" }
+clap = {version="4.1.6", features=["derive", "cargo", "unicode"]}
 #lz4_flex = "0.9.3"
 
 # [target.'cfg(unix)'.dependencies]
